refactor: changing how the source code is organized.
deleting a few unecessary dependencies and adding the Google Arc
Components.
