+++ PopSeries/app/src/main/AndroidManifest.xml
@@ -11,7 +11,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".home.HomeActivity">
+        <activity android:name=".ui.HomeActivity"
+            android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -20,7 +21,7 @@
         </activity>
         <activity
             android:name=".ui.DetailsActivity"
-            android:parentActivityName=".home.HomeActivity"/>
+            android:parentActivityName=".ui.HomeActivity"/>
     </application>
 
 </manifest>
