+++ PopSeries/app/src/main/java/alifyz/com/popseries/network/NetworkHelper.kt
@@ -1,3 +1,3 @@
 package alifyz.com.popseries.network
 
-//TODO - Implement the Retrofit Instance
+//TODO - Implement the RetrofitHelper Instance
